requirements.txt
opencv-contrib-python
numpy
pillow
pandas

# head to readme for more information